requirements.txt
google-generativeai>=0.8.0
google-api-python-client
google-auth